Clean up API unit tests for action executions
Split up the unit tests from test_executions_simple to multiple files.
